Include simple echo AWS Lambda implementation (#4)
This commit introduces support for a Kotlin/Multiplatoform, and by
extension Kotlin/Native, AWS Lambda Runtime Interface Client.
Currently this is a partial implementation that I expect will mature
as we use it.

Climate Changemakers will use this slack bot infrequently to submit
Hour of Action attendance data. Kotlin/Native is important here for
keeping costs to an absolute minimum. The cold start times of a
Kotlin/Native binary compared to a JVM AWS Lambda are substantial.

Plus, this was an interesting project.

To try this simple echo AWS Lambda function, you may use the test
docker image built from test.dockerfile.

```bash
$ docker build -t hoa-slackbot:latest . -f test.dockerfile
$ docker run -p 9000:8080 hoa-slackbot:latest
```

This will run the AWS Runtime Interface Emulator as part of the
docker image. This is a lightweight web server which emulates
AWS Lambda and allows you to test your functions locally.

Closes #1, #2